BAUDRATE from int to long.

+++ examples/TMC429_TMC26X_EVAL/TMC429_TMC26X_EVAL.ino
@@ -5,7 +5,7 @@
 #include "TMC26X.h"
 
 
-const int BAUDRATE = 115200;
+const long BAUDRATE = 115200;
 const int LOOP_DELAY = 1000;
 const int CS_PIN_429 = 10;
 const int CLOCK_FREQUENCY_MHZ = 16;
